+++ apps/explorer/test/explorer/import_test.exs
@@ -1,0 +1,5 @@
+defmodule Explorer.Chain.ImportTest do
+  use Explorer.DataCase
+
+  doctest Explorer.Chain.Import
+end
